Created basic OneToOne relation for coach and group, cleaned up code

+++ Backend/werqout/src/main/java/com/werqout/werqout/models/User.java
@@ -13,7 +13,7 @@ public class User {
      */
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
+    private long id;
     /**
      * Username of user
      */
@@ -38,14 +38,14 @@ public class User {
      * @param email user's email
      * @param password  user's password
      */
-    public User(int id, String userName, String email, String password) {
+    public User(long id, String userName, String email, String password) {
         this.id = id;
         this.userName = userName;
         this.email = email;
         this.password = password;
     }
     //--------------------------------------------------- Getters and setters ---------------------------------------------------------------------------------
-    public int getId() {
+    public long getId() {
         return id;
     }
     public void setId(int id) {
